+++ src/pyTOPP/pyproject.toml
@@ -1,0 +1,25 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pyopenms-pytopp"
+version = "0.1.0"
+description = "pyTOPP command-line wrappers and helpers for pyOpenMS"
+readme = "README.md"
+requires-python = ">=3.9"
+license = { text = "BSD-3-Clause" }
+authors = [{ name = "OpenMS team", email = "webmaster@openms.de" }]
+dependencies = [
+  "pyopenms>=3",
+  "pyprophet>=3",
+  "CTDopts @ git+https://github.com/WorkflowConversion/CTDopts.git",
+]
+
+# expose CLI entry points for your tools
+[project.scripts]
+pyTOPP-PyProphet = "pyopenms.pytopp.tools.pytopp_pyprophet_cli:main"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
