+++ src/pyTOPP/src/pyopenms/pytopp/CTDsupport.py
@@ -456,75 +456,6 @@ def print_model_help(model, *, advanced: bool, cfg: HelpConfig | None = None) ->
             "\nTip: use --help-advanced (or --helphelp) for all options, including advanced."
         )
 
-
-def _is_nullish(v) -> bool:
-    # CTDopts uses a private sentinel (class name "_Null") for “not provided”
-    return v is None or type(v).__name__ in {"_Null", "Null"} or v == ""
-
-
-def _normalize_path_list(v):
-    if _is_nullish(v):
-        return []
-    if isinstance(v, (list, tuple)):
-        return [Path(str(x)) for x in v if not _is_nullish(x)]
-    return [Path(str(v))]
-
-
-def _normalize_path_one(v):
-    if _is_nullish(v):
-        return None
-    return Path(str(v))
-
-
-def parse_args_ctdopts(argv, model) -> dict:
-    """
-    Parse command line using CTDopts only (no CTDsupport).
-    - honors OpenMS-style flags: -ini, -write_ini, -write_param_ctd
-    - loads args from CTD, overrides with CLI, validates and casts
-    - returns a flat dict of values
-    """
-    # Accept OpenMS flags with '-' prefix and OpenMS names
-    directives = parse_cl_directives(
-        argv,
-        input_ctd="ini",
-        write_tool_ctd="write_ini",
-        write_param_ctd="write_param_ctd",
-        prefix="-",
-    )
-
-    # If -write_ini was requested, emit the tool CTD and exit
-    if directives.get("write_tool_ctd"):
-        out = directives["write_tool_ctd"]
-        if out is True:
-            out = f"{getattr(model, 'name', 'tool')}.ctd"
-        model.write_ctd(out)  # tool schema only
-        print(f"Wrote tool CTD: {out}")
-        sys.exit(0)
-
-    # Seed args from -ini if given (values-only CTD)
-    base_args = {}
-    if directives.get("input_ctd"):
-        base_args = args_from_file(directives["input_ctd"])
-
-    # Parse CLI args (-prefixed, colon-separated for groups)
-    cli_args = model.parse_cl_args(argv, prefix="-")
-
-    # Merge and validate (CLI overrides file defaults)
-    merged = override_args(base_args, cli_args)
-    validated = model.validate_args(merged)
-
-    # If -write_param_ctd was requested, write current values and exit
-    if directives.get("write_param_ctd"):
-        out = directives["write_param_ctd"]
-        if out is True:
-            out = f"{getattr(model, 'name', 'tool')}_params.ctd"
-        model.write_ctd(out, validated)
-        print(f"Wrote parameter CTD: {out}")
-        sys.exit(0)
-
-    return validated
-
-
 # code related to CTD support
 def addParamToCTDopts(defaults, model):
     keys = defaults.keys()
