+++ src/pyTOPP/src/pyopenms/pytopp/tools/MapAlignerPoseClustering.py
